Return the correct error details when handling findById responses
According to the XHR docs[1], we can rely upon the presence of .status
and .statusText properties when handling XHR responses. As today's code
stands, if a findById request encounters an error, a blanket 'not found'
error is returned, ignoring the real HTTP status code and text. This
code makes those status code and status text available to the findById
callback.

[1] https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest