Add options argument to fetchById method; Use options in getOrFetch
Previously, `fetchById` only accepted `id` and `cb`, missing the `options` argument which is present in most, if not all ampersand-sync methods.

As a result, `getOrFetch` allowed the ability to specify an `options` object, but that was only used when retrieving the entire collection (with `options.all`). Now that `fetchById` accepts an options argument, `getOrFetch` can use the `options` argument when fetching just a single model.

Also, used ampersand-sync's `always` method instead of assigning a wrapped callback to both `success` and `error`